refactor(openomni): single canonical inbound payload-text parser
ingress extractPrompt (handlers.ts) and dispatch extractText
(dispatch/handlers/shared.ts) were the same payload-shape logic
maintained twice. Ingress owns it (the payload shape is minted at the
ingress boundary): the canonical extractText now lives in
ingress/handlers.ts and dispatch imports it; shared.ts keeps asRecord
only.

The two copies differed solely on nullish payloads — extractPrompt
returned JSON.stringify output ('null' for null, and a type-lying
runtime undefined for undefined) where extractText fails safe to .
The canonical takes the fail-safe body: the dispatch behavior is
test-pinned (extractText returns empty string for nullish payloads),
no test pins the ingress 'null'-prompt artifact, and a literal 'null'
prompt for a null payload was a latent bug, not a contract.
